SD card fix

+++ Core/Inc/liquid_sensor.h
@@ -18,6 +18,6 @@
 
 void liquid_sensor_begin();
 uint16_t get_liquid_adc();
-float get_liquid_liters();
+int32_t get_liquid_liters();
 
 #endif /* INC_LIQUID_SENSOR_H_ */
